Recognition/requirements.txt
Flask==2.0.1
PyTorch==1.10.0
torchvision==0.11.1
numpy==1.21.2
Pillow==8.3.2
matplotlib==3.4.3
scikit-learn==0.24.2
pandas==1.3.3
opencv-python==4.5.3.56
tqdm==4.62.3